ECON/ECON/requirements.txt
# ECON Framework Dependencies
# Core frameworks
torch>=2.0.0
numpy>=1.21.0
datasets>=2.14.0
transformers>=4.30.0

# Logging and monitoring
loguru>=0.7.0
wandb>=0.15.0
tensorboard>=2.12.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Environment and interaction
gymnasium>=0.28.0
gym>=0.26.0

# API and HTTP
requests>=2.31.0
openai>=1.0.0
together>=0.2.0

# Data processing
pandas>=2.0.0 